pyproject.toml
[project]
name = "pydpm_xl"
version = "0.1.39rc32"
description = "Python library for DPM-XL data processing and analysis"
authors = [
    {name = "MeaningfulData S.L.", email = "info@meaningfuldata.eu"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "GPL-3.0-or-later"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Database",
    "Topic :: Scientific/Engineering"
]
keywords = ["dpm", "dpm-xl", "data-processing", "migration", "analysis"]
dependencies = [
    "rich>=13.7.1,<13.8.0",
    "sqlalchemy>=1.4.50,<1.5.0",
    "pandas>=2.1.4",
    "antlr4-python3-runtime>=4.9.2,<4.9.3",
    "pyodbc>=5.1.0,<5.2.0",
    "click>=8.1.6,<8.2.0",
    "python-dotenv>=1.0.1,<1.1.0",
    "psycopg2-binary>=2.9.0,<3.0.0"
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0,<8.0.0",
    "pytest-cov>=4.1.0,<5.0.0"
]

[project.scripts]
pydpm = "py_dpm.cli.main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["py_dpm*"]
exclude = []

[tool.poetry]
name = "pydpm_xl"
version = "0.1.39rc32"
description = "Python library for DPM-XL data processing and analysis"
authors = ["MeaningfulData S.L. <info@meaningfuldata.eu>"]
readme = "README.md"
license = "GPL-3.0-or-later"
packages = [{include = "py_dpm"}]

[tool.poetry.dependencies]
python = ">=3.10"
rich = ">=13.7.1,<13.8.0"
sqlalchemy = ">=1.4.50,<1.5.0"
pandas = ">=2.1.4"
antlr4-python3-runtime = ">=4.9.2,<4.9.3"
pyodbc = ">=5.1.0,<5.2.0"
click = ">=8.1.6,<8.2.0"
python-dotenv = ">=1.0.1,<1.1.0"
psycopg2-binary = ">=2.9.0,<3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0,<8.0.0"
pytest-cov = ">=4.1.0,<5.0.0"

[tool.poetry.scripts]
pydpm = "py_dpm.cli.main:main"

[tool.setuptools.package-data]
"py_dpm.dpm_xl.grammar" = ["*.g4"]
"py_dpm.dpm_xl.grammar.generated" = ["*"]
"py_dpm.dpm.db.views" = ["*.sql"]
